+++ src/os2datascanner/projects/report/tests/test_msgraph_utilities.py
@@ -117,13 +117,11 @@ def msgraph_mail_reports():  # noqa: ECE001
 
     return [
         DocumentReport(
-            name="Report-nested",
             owner="jean@claude.sky",
             path="docx",
             raw_metadata=flat_metadata
         ),
         DocumentReport(
-            name="Report-nested",
             owner="jean@claude.sky",
             path="docx",
             raw_metadata=nested_metadata
@@ -161,7 +159,6 @@ def not_msgraph_reports():
 
     filescan_metadata.update(scan_tag_origin_metadata)
     return DocumentReport(
-        name="Report-nested",
         owner="jean@claude.sky",
         path="docx",
         raw_metadata=filescan_metadata,